+++ include/emthrm/graph/eulerian_trail_in_directed_graph.hpp
@@ -59,7 +59,7 @@ std::vector<Edge<CostType>> eulerian_trail_in_directed_graph(
     }
   };
   dfs(s);
-  if (res.size() == edge_num) {
+  if (static_cast<int>(res.size()) == edge_num) {
     std::reverse(res.begin(), res.end());
     return res;
   }
