+++ include/emthrm/graph/flow/minimum_cost_flow/minimum_cost_b-flow.hpp
@@ -70,7 +70,7 @@ struct MinimumCostBFlow {
         std::tie(d, ver) = que.top();
         que.pop();
         if (d > dist[ver]) continue;
-        for (int i = 0; i < graph[ver].size(); ++i) {
+        for (int i = 0; i < static_cast<int>(graph[ver].size()); ++i) {
           const Edge& e = graph[ver][i];
           const U nxt = dist[ver] + e.cost + potential[ver] - potential[e.dst];
           if (e.cap > 0 && dist[e.dst] > nxt) {
