+++ include/emthrm/math/modint.hpp
@@ -91,11 +91,11 @@
 //     return res;
 //   }
 //   MInt& operator+=(const MInt& x) {
-//     if ((v += x.v) >= M) v -= M;
+//     if (static_cast<int>(v += x.v) >= M) v -= M;
 //     return *this;
 //   }
 //   MInt& operator-=(const MInt& x) {
-//     if ((v += M - x.v) >= M) v -= M;
+//     if (static_cast<int>(v += M - x.v) >= M) v -= M;
 //     return *this;
 //   }
 //   MInt& operator*=(const MInt& x) {
@@ -110,7 +110,7 @@
 //   bool operator>(const MInt& x) const { return v > x.v; }
 //   bool operator>=(const MInt& x) const { return v >= x.v; }
 //   MInt& operator++() {
-//     if (++v == M) v = 0;
+//     if (static_cast<int>(++v) == M) v = 0;
 //     return *this;
 //   }
 //   MInt operator++(int) {
@@ -230,11 +230,11 @@ struct MInt {
     return res;
   }
   MInt& operator+=(const MInt& x) {
-    if ((v += x.v) >= mod()) v -= mod();
+    if (static_cast<int>(v += x.v) >= mod()) v -= mod();
     return *this;
   }
   MInt& operator-=(const MInt& x) {
-    if ((v += mod() - x.v) >= mod()) v -= mod();
+    if (static_cast<int>(v += mod() - x.v) >= mod()) v -= mod();
     return *this;
   }
   MInt& operator*=(const MInt& x) {
@@ -249,7 +249,7 @@ struct MInt {
   bool operator>(const MInt& x) const { return v > x.v; }
   bool operator>=(const MInt& x) const { return v >= x.v; }
   MInt& operator++() {
-    if (++v == mod()) v = 0;
+    if (static_cast<int>(++v) == mod()) v = 0;
     return *this;
   }
   MInt operator++(int) {
