+++ test/graph/flow/minimum_cost_flow/minimum_cost_s-t-flow.3.test.cpp
@@ -37,7 +37,7 @@ int main() {
     }
     MinimumCostSTFlow<int, long long> minimum_cost_flow(num + 2);
     const int s = num, t = num + 1;
-    for (int i = 0; i < trains.front().size(); ++i) {
+    for (int i = 0; i < static_cast<int>(trains.front().size()); ++i) {
       minimum_cost_flow.add_edge(s, i, 1, 0);
     }
     int w = 0;
@@ -57,7 +57,7 @@ int main() {
       w += m;
       if (i + 1 < n - 1) {
         for (int j = 0; j < m; ++j) {
-          for (int k = 0; k < trains[i + 1].size(); ++k) {
+          for (int k = 0; k < static_cast<int>(trains[i + 1].size()); ++k) {
             if (times[i][j] <= trains[i + 1][k].x) {
               minimum_cost_flow.add_edge(j + w, k + w + m, 1, 0);
             }
