+++ test/string/split.test.cpp
@@ -16,7 +16,7 @@ int main() {
   int ans = 0;
   for (const std::string& terms : split(s, "+")) {
     bool is_0 = false;
-    for (int i = 0; i < terms.length(); i += 2) {
+    for (int i = 0; i < static_cast<int>(terms.length()); i += 2) {
       is_0 |= terms[i] == '0';
     }
     ans += !is_0;
