+++ test/string/z_algorithm.test.cpp
@@ -13,8 +13,8 @@ int main() {
   std::string s;
   std::cin >> s;
   const std::vector<int> ans = z_algorithm(s);
-  for (int i = 0; i < s.length(); ++i) {
-    std::cout << ans[i] << " \n"[i + 1 == s.length()];
+  for (int i = 0; i < static_cast<int>(s.length()); ++i) {
+    std::cout << ans[i] << " \n"[i + 1 == static_cast<int>(s.length())];
   }
   return 0;
 }
